+++ packages/bluesky/lib/src/services/codegen/tools/ozone/queue/defs/queue_view.g.dart
@@ -32,6 +32,10 @@ _QueueView _$QueueViewFromJson(Map json) => $checkedCreate('_QueueView', json, (
       (v) => (v as List<dynamic>?)?.map((e) => e as String).toList(),
     ),
     description: $checkedConvert('description', (v) => v as String?),
+    recommendedPolicies: $checkedConvert(
+      'recommendedPolicies',
+      (v) => (v as List<dynamic>?)?.map((e) => e as String).toList(),
+    ),
     createdBy: $checkedConvert('createdBy', (v) => v as String),
     createdAt: $checkedConvert('createdAt', (v) => DateTime.parse(v as String)),
     updatedAt: $checkedConvert('updatedAt', (v) => DateTime.parse(v as String)),
@@ -63,6 +67,7 @@ Map<String, dynamic> _$QueueViewToJson(_QueueView instance) =>
       'collection': ?instance.collection,
       'reportTypes': ?instance.reportTypes,
       'description': ?instance.description,
+      'recommendedPolicies': ?instance.recommendedPolicies,
       'createdBy': instance.createdBy,
       'createdAt': iso8601(instance.createdAt),
       'updatedAt': iso8601(instance.updatedAt),