+++ packages/bluesky/lib/src/services/codegen/tools/ozone/queue/updateQueue/input.freezed.dart
@@ -19,7 +19,7 @@ mixin _$QueueUpdateQueueInput {
  int get queueId;/// New display name for the queue
  String? get name;/// Enable or disable the queue
  bool? get enabled;/// Optional description of the queue
- String? get description; Map<String, dynamic>? get $unknown;
+ String? get description; List<String>? get recommendedPolicies; Map<String, dynamic>? get $unknown;
 /// Create a copy of QueueUpdateQueueInput
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -32,16 +32,16 @@ $QueueUpdateQueueInputCopyWith<QueueUpdateQueueInput> get copyWith => _$QueueUpd
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is QueueUpdateQueueInput&&(identical(other.queueId, queueId) || other.queueId == queueId)&&(identical(other.name, name) || other.name == name)&&(identical(other.enabled, enabled) || other.enabled == enabled)&&(identical(other.description, description) || other.description == description)&&const DeepCollectionEquality().equals(other.$unknown, $unknown));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is QueueUpdateQueueInput&&(identical(other.queueId, queueId) || other.queueId == queueId)&&(identical(other.name, name) || other.name == name)&&(identical(other.enabled, enabled) || other.enabled == enabled)&&(identical(other.description, description) || other.description == description)&&const DeepCollectionEquality().equals(other.recommendedPolicies, recommendedPolicies)&&const DeepCollectionEquality().equals(other.$unknown, $unknown));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,queueId,name,enabled,description,const DeepCollectionEquality().hash($unknown));
+int get hashCode => Object.hash(runtimeType,queueId,name,enabled,description,const DeepCollectionEquality().hash(recommendedPolicies),const DeepCollectionEquality().hash($unknown));
 
 @override
 String toString() {
-  return 'QueueUpdateQueueInput(queueId: $queueId, name: $name, enabled: $enabled, description: $description, \$unknown: ${$unknown})';
+  return 'QueueUpdateQueueInput(queueId: $queueId, name: $name, enabled: $enabled, description: $description, recommendedPolicies: $recommendedPolicies, \$unknown: ${$unknown})';
 }
 
 
@@ -52,7 +52,7 @@ abstract mixin class $QueueUpdateQueueInputCopyWith<$Res>  {
   factory $QueueUpdateQueueInputCopyWith(QueueUpdateQueueInput value, $Res Function(QueueUpdateQueueInput) _then) = _$QueueUpdateQueueInputCopyWithImpl;
 @useResult
 $Res call({
- int queueId, String? name, bool? enabled, String? description, Map<String, dynamic>? $unknown
+ int queueId, String? name, bool? enabled, String? description, List<String>? recommendedPolicies, Map<String, dynamic>? $unknown
 });
 
 
@@ -69,13 +69,14 @@ class _$QueueUpdateQueueInputCopyWithImpl<$Res>
 
 /// Create a copy of QueueUpdateQueueInput
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? queueId = null,Object? name = freezed,Object? enabled = freezed,Object? description = freezed,Object? $unknown = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? queueId = null,Object? name = freezed,Object? enabled = freezed,Object? description = freezed,Object? recommendedPolicies = freezed,Object? $unknown = freezed,}) {
   return _then(_self.copyWith(
 queueId: null == queueId ? _self.queueId : queueId // ignore: cast_nullable_to_non_nullable
 as int,name: freezed == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
 as String?,enabled: freezed == enabled ? _self.enabled : enabled // ignore: cast_nullable_to_non_nullable
 as bool?,description: freezed == description ? _self.description : description // ignore: cast_nullable_to_non_nullable
-as String?,$unknown: freezed == $unknown ? _self.$unknown : $unknown // ignore: cast_nullable_to_non_nullable
+as String?,recommendedPolicies: freezed == recommendedPolicies ? _self.recommendedPolicies : recommendedPolicies // ignore: cast_nullable_to_non_nullable
+as List<String>?,$unknown: freezed == $unknown ? _self.$unknown : $unknown // ignore: cast_nullable_to_non_nullable
 as Map<String, dynamic>?,
   ));
 }
@@ -161,10 +162,10 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( int queueId,  String? name,  bool? enabled,  String? description,  Map<String, dynamic>? $unknown)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( int queueId,  String? name,  bool? enabled,  String? description,  List<String>? recommendedPolicies,  Map<String, dynamic>? $unknown)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _QueueUpdateQueueInput() when $default != null:
-return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.$unknown);case _:
+return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.recommendedPolicies,_that.$unknown);case _:
   return orElse();
 
 }
@@ -182,10 +183,10 @@ return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.$
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( int queueId,  String? name,  bool? enabled,  String? description,  Map<String, dynamic>? $unknown)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( int queueId,  String? name,  bool? enabled,  String? description,  List<String>? recommendedPolicies,  Map<String, dynamic>? $unknown)  $default,) {final _that = this;
 switch (_that) {
 case _QueueUpdateQueueInput():
-return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.$unknown);case _:
+return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.recommendedPolicies,_that.$unknown);case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -202,10 +203,10 @@ return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.$
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( int queueId,  String? name,  bool? enabled,  String? description,  Map<String, dynamic>? $unknown)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( int queueId,  String? name,  bool? enabled,  String? description,  List<String>? recommendedPolicies,  Map<String, dynamic>? $unknown)?  $default,) {final _that = this;
 switch (_that) {
 case _QueueUpdateQueueInput() when $default != null:
-return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.$unknown);case _:
+return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.recommendedPolicies,_that.$unknown);case _:
   return null;
 
 }
@@ -217,7 +218,7 @@ return $default(_that.queueId,_that.name,_that.enabled,_that.description,_that.$
 
 @JsonSerializable(includeIfNull: false)
 class _QueueUpdateQueueInput implements QueueUpdateQueueInput {
-  const _QueueUpdateQueueInput({required this.queueId, this.name, this.enabled, this.description, final  Map<String, dynamic>? $unknown}): _$unknown = $unknown;
+  const _QueueUpdateQueueInput({required this.queueId, this.name, this.enabled, this.description, final  List<String>? recommendedPolicies, final  Map<String, dynamic>? $unknown}): _recommendedPolicies = recommendedPolicies,_$unknown = $unknown;
   factory _QueueUpdateQueueInput.fromJson(Map<String, dynamic> json) => _$QueueUpdateQueueInputFromJson(json);
 
 /// ID of the queue to update
@@ -228,6 +229,15 @@ class _QueueUpdateQueueInput implements QueueUpdateQueueInput {
 @override final  bool? enabled;
 /// Optional description of the queue
 @override final  String? description;
+ final  List<String>? _recommendedPolicies;
+@override List<String>? get recommendedPolicies {
+  final value = _recommendedPolicies;
+  if (value == null) return null;
+  if (_recommendedPolicies is EqualUnmodifiableListView) return _recommendedPolicies;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableListView(value);
+}
+
  final  Map<String, dynamic>? _$unknown;
 @override Map<String, dynamic>? get $unknown {
   final value = _$unknown;
@@ -251,16 +261,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _QueueUpdateQueueInput&&(identical(other.queueId, queueId) || other.queueId == queueId)&&(identical(other.name, name) || other.name == name)&&(identical(other.enabled, enabled) || other.enabled == enabled)&&(identical(other.description, description) || other.description == description)&&const DeepCollectionEquality().equals(other._$unknown, _$unknown));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _QueueUpdateQueueInput&&(identical(other.queueId, queueId) || other.queueId == queueId)&&(identical(other.name, name) || other.name == name)&&(identical(other.enabled, enabled) || other.enabled == enabled)&&(identical(other.description, description) || other.description == description)&&const DeepCollectionEquality().equals(other._recommendedPolicies, _recommendedPolicies)&&const DeepCollectionEquality().equals(other._$unknown, _$unknown));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,queueId,name,enabled,description,const DeepCollectionEquality().hash(_$unknown));
+int get hashCode => Object.hash(runtimeType,queueId,name,enabled,description,const DeepCollectionEquality().hash(_recommendedPolicies),const DeepCollectionEquality().hash(_$unknown));
 
 @override
 String toString() {
-  return 'QueueUpdateQueueInput(queueId: $queueId, name: $name, enabled: $enabled, description: $description, \$unknown: ${$unknown})';
+  return 'QueueUpdateQueueInput(queueId: $queueId, name: $name, enabled: $enabled, description: $description, recommendedPolicies: $recommendedPolicies, \$unknown: ${$unknown})';
 }
 
 
@@ -271,7 +281,7 @@ abstract mixin class _$QueueUpdateQueueInputCopyWith<$Res> implements $QueueUpda
   factory _$QueueUpdateQueueInputCopyWith(_QueueUpdateQueueInput value, $Res Function(_QueueUpdateQueueInput) _then) = __$QueueUpdateQueueInputCopyWithImpl;
 @override @useResult
 $Res call({
- int queueId, String? name, bool? enabled, String? description, Map<String, dynamic>? $unknown
+ int queueId, String? name, bool? enabled, String? description, List<String>? recommendedPolicies, Map<String, dynamic>? $unknown
 });
 
 
@@ -288,13 +298,14 @@ class __$QueueUpdateQueueInputCopyWithImpl<$Res>
 
 /// Create a copy of QueueUpdateQueueInput
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? queueId = null,Object? name = freezed,Object? enabled = freezed,Object? description = freezed,Object? $unknown = freezed,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? queueId = null,Object? name = freezed,Object? enabled = freezed,Object? description = freezed,Object? recommendedPolicies = freezed,Object? $unknown = freezed,}) {
   return _then(_QueueUpdateQueueInput(
 queueId: null == queueId ? _self.queueId : queueId // ignore: cast_nullable_to_non_nullable
 as int,name: freezed == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
 as String?,enabled: freezed == enabled ? _self.enabled : enabled // ignore: cast_nullable_to_non_nullable
 as bool?,description: freezed == description ? _self.description : description // ignore: cast_nullable_to_non_nullable
-as String?,$unknown: freezed == $unknown ? _self._$unknown : $unknown // ignore: cast_nullable_to_non_nullable
+as String?,recommendedPolicies: freezed == recommendedPolicies ? _self._recommendedPolicies : recommendedPolicies // ignore: cast_nullable_to_non_nullable
+as List<String>?,$unknown: freezed == $unknown ? _self._$unknown : $unknown // ignore: cast_nullable_to_non_nullable
 as Map<String, dynamic>?,
   ));
 }